requirements.txt
streamlit
tensorflow==2.17.1
pillow
numpy
streamlit_webrtc